+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Improved
+- **perf**: Prevent need for async regex behaviors in Script by adding a __name shim (needed because bundlers like esbuild/vite/... work like this) in atomic runtime.
+
 ## [0.36.1] - 2025-06-21
 ### Improved
 - **dx**: Adjust typing for AppOptions `client.css` to be of `CssInstance`